+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ### Added
 - Settings to change transparency values.
+- Settings to change default note size.
 - 
 
 ### Changed